Remove metadata in favor of custom fields (#420)
## Description

<!-- Provide a brief description of the PR's purpose here. -->

We plan to add the ability to specify custom fields in pyribs archives.
This PR thus removes metadata, which are a current substitute for custom
fields. Custom fields will also support object dtypes, meaning metadata
can still be recovered by specifying a metadata field with object dtype.

While this change is backwards-incompatible, I believe it is beneficial
for the simplicity of the library. Currently, metadata is the only array
that is an object array, making it more difficult to save next to arrays
like the objective and measures arrays. Furthermore, metadata currently
has a complex defaulting system since it is usually None.

## TODO

<!-- Notable points that this PR has either accomplished or will
accomplish. -->

- [x] Remove metadata from codebase
- [x] Remove metadata from tests
- [x] Fix docstrings

## Questions

<!-- Any concerns or points of confusion? -->

## Status

- [x] I have read the guidelines in

[CONTRIBUTING.md](https://github.com/icaros-usc/pyribs/blob/master/CONTRIBUTING.md)
- [x] I have formatted my code using `yapf`
- [x] I have tested my code by running `pytest`
- [x] I have linted my code with `pylint`
- [x] I have added a one-line description of my change to the changelog
in
      `HISTORY.md`
- [x] This PR is ready to go