Add ranking values to evolution strategy tell method (#438)
## Description

<!-- Provide a brief description of the PR's purpose here. -->

This PR adds the `ranking_values` argument to the tell() method in
evolution strategies. This paves the way for #434, since pycma needs the
ranking values to be passed into tell().

Note that I considered removing ranking_values from `check_stop` and
instead forcing the ES to store its own ranking values like pycma does,
but I do not think this is a necessary change.

This change is backwards-incompatible, but similar to #436, I do not
think it will affect most users.

## TODO

<!-- Notable points that this PR has either accomplished or will
accomplish. -->

- [x] Update EvolutionStrategyBase and corresponding ESs
- [x] Update emitter calls
- [x] Remove redundant method docstrings

## Questions

<!-- Any concerns or points of confusion? -->

## Status

- [x] I have read the guidelines in

[CONTRIBUTING.md](https://github.com/icaros-usc/pyribs/blob/master/CONTRIBUTING.md)
- [x] I have formatted my code using `yapf`
- [x] I have tested my code by running `pytest`
- [x] I have linted my code with `pylint`
- [x] I have added a one-line description of my change to the changelog
in
      `HISTORY.md`
- [x] This PR is ready to go