Support seed=SeedSequence in emitters (#440)
## Description

<!-- Provide a brief description of the PR's purpose here. -->

Previously, passing a SeedSequence as a seed in EvolutionStrategyEmitter
and GradientArborescenceEmitter threw an error because SeedSequence
cannot be initialized from another SeedSequence. This PR adds checks for
SeedSequence so that SeedSequence can be passed in as a seed in these
two emitters. Note that in other emitters, the seed is passed directly
to the default_rng(), which can handle SeedSequence as input.

## Status

- [x] I have read the guidelines in

[CONTRIBUTING.md](https://github.com/icaros-usc/pyribs/blob/master/CONTRIBUTING.md)
- [x] I have formatted my code using `yapf`
- [x] I have tested my code by running `pytest`
- [x] I have linted my code with `pylint`
- [x] I have added a one-line description of my change to the changelog
in
      `HISTORY.md`
- [x] This PR is ready to go