Raise error if `result_archive` and `archive` have different fields (#461)
## Description

<!-- Provide a brief description of the PR's purpose here. -->

In the scheduler, we did not previously check that result_archive and
archive had the same fields. This could lead to confusing errors where
we insert certain data into the archive, but it fails to insert into the
result_archive. This PR adds a check so that this does not happen. It
also fixes some other scheduler documentation.

## Status

- [x] I have read the guidelines in

[CONTRIBUTING.md](https://github.com/icaros-usc/pyribs/blob/master/CONTRIBUTING.md)
- [x] I have formatted my code using `yapf`
- [x] I have tested my code by running `pytest`
- [x] I have linted my code with `pylint`
- [x] I have added a one-line description of my change to the changelog
in
      `HISTORY.md`
- [x] This PR is ready to go