Replace archive.dtype with archive.dtypes dict that holds dtype of every field (#470)
## Description

<!-- Provide a brief description of the PR's purpose here. -->

Previously, the archives had a `dtype` property that held the dtype of
the solution, objectives, and measures. However, these three may not
always have the same dtype; furthermore, this excludes the threshold
field and other additional fields provided by the user. Thus, this PR
replaces `dtype` with `dtypes`, a dict that maps from field name to
dtype of the associated field.

## TODO

<!-- Notable points that this PR has either accomplished or will
accomplish. -->

- [x] Deprecate `ArchiveBase.dtype` -- now, calling archive.dtype will
raise a RuntimeError with a deprecation notice
- [x] Add `ArchvieBase.dtypes`
- [x] Update calls to `archive.dtype`. In particular, we now use
`archive.dtypes["XXXX"]`, with XXXX being the field name that is
appropriate to the use case, e.g., `x0` in an emitter would be cast to
`archive.dtypes["solution"]`.
- [x] Note that it was previously common to call
`archive.dtype(some_scalar_value)`. This is not very correct. To
elaborate, `archive.dtype` was a numpy scalar type, e.g., `np.float32`
or `np.float64`, which is indeed callable. However, the actual dtype of
an array, e.g., `np.arange(5).dtype` is an instance of `numpy.dtype`,
which is _not_ callable. Thus, a utility called `np_scalar` was
introduced, which casts a provided value to a provided dtype, e.g.,
`np_scalar(5.0, dtype=archive.dtypes["solution"])`

## Status

- [x] I have read the guidelines in

[CONTRIBUTING.md](https://github.com/icaros-usc/pyribs/blob/master/CONTRIBUTING.md)
- [x] I have formatted my code using `yapf`
- [x] I have tested my code by running `pytest`
- [x] I have linted my code with `pylint`
- [x] I have added a one-line description of my change to the changelog
in
      `HISTORY.md`
- [x] This PR is ready to go