Support novelty search with local competition in ProximityArchive (#481)
## Description

<!-- Provide a brief description of the PR's purpose here. -->

Support NSLC in the ProximityArchive. This involves modifying the `add`
method in particular to compute local competition as well as value. It
should be noted that the NSLC paper does not specify how replacement
should be handled in the archive -- it would seem that the authors only
use a regular novelty search archive. We have decided to implement a
rule where if the solution is not novel enough for addition, it shall be
compared to its nearest neighbor and added if it exceeds the objective
value of its nearest neighbor.

## TODO

<!-- Notable points that this PR has either accomplished or will
accomplish. -->

- [x] Add `local_competition` flag to ProximityArchive
- [x] Add tests

## Status

- [x] I have read the guidelines in

[CONTRIBUTING.md](https://github.com/icaros-usc/pyribs/blob/master/CONTRIBUTING.md)
- [x] I have formatted my code using `yapf`
- [x] I have tested my code by running `pytest`
- [x] I have linted my code with `pylint`
- [x] I have added a one-line description of my change to the changelog
in
      `HISTORY.md`
- [x] This PR is ready to go