Move novelty computation into a separate method (#484)
## Description

<!-- Provide a brief description of the PR's purpose here. -->

Adds a new public `compute_novelty` method to `ProximityArchive` to
allow computing the novelty and local competition scores of arbitrary
measures, not just of solutions that are passed into `add`. Previously,
the operations in this method were part of `add`.

Regarding API, I allow the `local_competition` argument to either be
None or an array of objectives to use during the local competition
computation. I also considered an API that passed in objectives
separately to be a bit more consistent with the `add` API, eg,
`compute_novelty(objective, measures, local_competition)` but it seemed
clunky to always require calling `compute_novelty(None, measures)`.

## Status

- [x] I have read the guidelines in

[CONTRIBUTING.md](https://github.com/icaros-usc/pyribs/blob/master/CONTRIBUTING.md)
- [x] I have formatted my code using `yapf`
- [x] I have tested my code by running `pytest`
- [x] I have linted my code with `pylint`
- [x] I have added a one-line description of my change to the changelog
in
      `HISTORY.md`
- [x] This PR is ready to go