BanditScheduler: Add emitter_pool and active attr; remove emitters attr (#494)
## Description

<!-- Provide a brief description of the PR's purpose here. -->

Currently, the BanditScheduler has an `emitters` attr that returns the
`_active_arr` variable. This does not make much sense since
BanditScheduler does not have `emitters` as input, and `_active_arr`
gives `emitters` a different meaning from the `emitters` attribute of
Scheduler.

Thus, this PR removes the `emitters` attr and instead adds
`emitter_pool` and `active` to indicate the pool of emitters in the
scheduler and which ones are currently active, respectively.

## TODO

<!-- Notable points that this PR has either accomplished or will
accomplish. -->

- [x] Remove `BanditScheduler.emitters`
- [x] Add `BanditScheduler.emitter_pool`
- [x] Add `BanditScheduler.active`

## Status

- [x] I have read the guidelines in

[CONTRIBUTING.md](https://github.com/icaros-usc/pyribs/blob/master/CONTRIBUTING.md)
- [x] I have formatted my code using `yapf`
- [x] I have tested my code by running `pytest`
- [x] I have linted my code with `pylint`
- [x] I have added a one-line description of my change to the changelog
in
      `HISTORY.md`
- [x] This PR is ready to go