Make DensityRanker use add_info (#504)
## Description

<!-- Provide a brief description of the PR's purpose here. -->

Previously, `DensityRanker` called `compute_density` on the
(soon-to-be-implemented) `DensityArchive`.

Here are some considerations -- there are two places where we can call
`compute_density`, either in the ranker, or in `DensityArchive.add`:

- If we have 15 emitters with 15 rankers, each ranker will call
`compute_density` on its own, which seems inefficient. If we put
`compute_density` in `add`, we only call it once.
- If we call `compute_density` on its own, we do get the most up-to-date
density information, but I don't foresee the archive changing in between
addition of solutions and the emitters updating their rankings.
- We keep the rankers simple since we keep the call to `compute_density`
out of them.
- `ProximityArchive` computes novelty in `add`, so calling in `add`
would maintain consistency with `ProximityArchive`.

Overall, to maintain consistency and also make the calls be centralized,
I think this change makes sense.

## Status

- [x] I have read the guidelines in

[CONTRIBUTING.md](https://github.com/icaros-usc/pyribs/blob/master/CONTRIBUTING.md)
- [x] I have formatted my code using `yapf`
- [x] I have tested my code by running `pytest`
- [x] I have linted my code with `pylint`
- [x] I have added a one-line description of my change to the changelog
in
      `HISTORY.md`
- [x] This PR is ready to go