Set vmin and vmax to None if archive is empty in ribs.visualize (#513)
## Description

<!-- Provide a brief description of the PR's purpose here. -->

Currently, if an empty archive is passed to the visualization tools and
`vmin` and `vmax` are not set, we will throw an error because we try to
take `np.min` or `np.max` of an empty array. There is also an instance
in `archive_heatmap_1d` where we call `np.nanmin` and `np.nanmax` on an
all-NaN array, which raises a warning.

This PR checks that `vmin` and `vmax` are None, inf, or nan when the
archive is empty, so that we use the default bounds available in
Matplotlib.

## TODO

<!-- Notable points that this PR has either accomplished or will
accomplish. -->

- [x] cvt_archive_3d_plot -> following cvt_archive_heatmap, vmin and
vmax default to -inf and inf so we get the default bounds
- [x] cvt_archive_heatmap -> vmin and vmax default to -inf and inf so we
get the default bounds
- [x] grid_archive_heatmap -> added checks for vmin and vmax; they now
set to None
- [x] parallel_axes_plot -> vmin and vmax get set to nan so we get the
default bounds
- [x] sliding_boundaries_archive_heatmap -> added checks for vmin and
vmax; they now set to None. Bounds end up being 0 to 1 instead of -0.1
to 0.1 because this function calls scatter rather than pcolormesh

Other things that came up:
- [x] Switched the `archive_heatmap_1d` function to take in only keyword
arguments (previously had a bunch of positional args)
- [x] Used `min_obj` and `max_obj` for scatterplot when `plot_elites` is
turned on in `cvt_archive_3d_plot`

## Status

- [x] I have read the guidelines in

[CONTRIBUTING.md](https://github.com/icaros-usc/pyribs/blob/master/CONTRIBUTING.md)
- [x] I have formatted my code using `yapf`
- [x] I have tested my code by running `pytest`
- [x] I have linted my code with `pylint`
- [x] I have added a one-line description of my change to the changelog
in
      `HISTORY.md`
- [x] This PR is ready to go