Increase max positional arguments for pylint (#506)
## Description

<!-- Provide a brief description of the PR's purpose here. -->

We tend to use a lot of positional arguments in our methods, like in
tell_dqd. There are also methods like `cqd_score` that don't have the
`*` to force keyword arguments. In some places like the sphere example,
it also doesn't make sense to add this `*` because it interferes with
fire. Thus, I increased the number of positional arguments allowed by
pylint.

## Status

- [x] I have read the guidelines in

[CONTRIBUTING.md](https://github.com/icaros-usc/pyribs/blob/master/CONTRIBUTING.md)
- [x] I have formatted my code using `yapf`
- [x] I have tested my code by running `pytest`
- [x] I have linted my code with `pylint`
- [N/A] I have added a one-line description of my change to the
changelog in
      `HISTORY.md`
- [x] This PR is ready to go