Fix bounds and limits in proximity_archive_plot (#523)
## Description

<!-- Provide a brief description of the PR's purpose here. -->

The bounds and limits need to have sensible default values when the
archive is empty. The limits were previously fixed for other
visualization functiions in #513, but this was accidentally left out.
The bounds also need default values since, unlike other archives,
ProximityArchive is unstructured and does not have bounds.

## TODO

- [x] Update tests -- add new baseline images too
- [x] Fix lower_bounds, upper_bounds, vmin, and vmax in
`proximity_archive_plot`

## Status

- [x] I have read the guidelines in

[CONTRIBUTING.md](https://github.com/icaros-usc/pyribs/blob/master/CONTRIBUTING.md)
- [x] I have formatted my code using `yapf`
- [x] I have tested my code by running `pytest`
- [x] I have linted my code with `pylint`
- [x] I have added a one-line description of my change to the changelog
in
      `HISTORY.md`
- [x] This PR is ready to go