Add retessellate method to GridArchive (#516)
## Description

<!-- Provide a brief description of the PR's purpose here. -->

This method changes the dimensionality of the grid in a `GridArchive`,
e.g., we can go from `10x10` to `20x20` or `5x5`. This method was
introduced by @szhaovas in #496. Here, I've split it out into a separate
PR that re-implements it as an in-place method on the archive, i.e., the
archive will be retessellated in-place.

## TODO

<!-- Notable points that this PR has either accomplished or will
accomplish. -->

- [x] Implement retessellate
- [x] Add tests

## Status

- [x] I have read the guidelines in

[CONTRIBUTING.md](https://github.com/icaros-usc/pyribs/blob/master/CONTRIBUTING.md)
- [x] I have formatted my code using `yapf`
- [x] I have tested my code by running `pytest`
- [x] I have linted my code with `pylint`
- [x] I have added a one-line description of my change to the changelog
in
      `HISTORY.md`
- [x] This PR is ready to go