Remove thresholds from ProximityArchive and SlidingBoundariesArchive (#527)
## Description

<!-- Provide a brief description of the PR's purpose here. -->

Currently, thresholds are included as a field in ProximityArchive and
SlidingBoundariesArchive, along with the corresponding properties
learning_rate and threshold_min. However, these archives are not
designed to support CMA-MAE in any way; rather, the threshold here is an
artifact of depending on the old ArchiveBase. Thus, this PR removes the
thresholds in these archives and updates the tests accordingly.

Furthermore, since some archives now have thresholds and some do not,
the field list check in Scheduler no longer works, since either
`archive` or `result_archive` may not have the `threshold` field (such
as in Novelty Search with `archive` as a ProximityArchive and
`result_archive` as a GridArchive). Hence, I removed the check and
instead clarified the error message in ArrayStore.

## TODO

<!-- Notable points that this PR has either accomplished or will
accomplish. -->

- [x] Remove thresholds from ProximityArchive 
- [x] Remove thresholds from SlidingBoundariesArchive
- [x] Update archive tests to not test for thresholds anymore
- [x] Remove checks for `field_list` in Scheduler and BanditScheduler
- [x] Update tests for `field_list` in `scheduler_test.py`
- [x] Update error message in ArrayStore to mention archive and
result_archive mismatch.

## Status

- [x] I have read the guidelines in

[CONTRIBUTING.md](https://github.com/icaros-usc/pyribs/blob/master/CONTRIBUTING.md)
- [x] I have formatted my code using `yapf`
- [x] I have tested my code by running `pytest`
- [x] I have linted my code with `pylint`
- [x] I have added a one-line description of my change to the changelog
in
      `HISTORY.md`
- [x] This PR is ready to go