Make archive field_list and dtypes props include index (#532)
## Description

<!-- Provide a brief description of the PR's purpose here. -->

In the past, `index` was assumed to be a part of every archive. However,
with the recent refactoring, `index` is no longer required to be in
archives; rather, it is specific to archives that use an `ArrayStore`.
Thus, I have updated the archives to include `index` in the list of
fields (`field_list`) and the dictionary of field data types (`dtypes`).

## TODO

<!-- Notable points that this PR has either accomplished or will
accomplish. -->

- [x] Add new `field_list_with_index` and `dtypes_with_index` props to
ArrayStore
- [x] Update archives' `field_list` to call `field_list_with_index`
- [x] Update archives' `dtypes` to call `dtypes_with_index`
- [x] Test ArrayStore
- [x] Test new field_list and dtypes in archives

## Status

- [x] I have read the guidelines in

[CONTRIBUTING.md](https://github.com/icaros-usc/pyribs/blob/master/CONTRIBUTING.md)
- [x] I have formatted my code using `yapf`
- [x] I have tested my code by running `pytest`
- [x] I have linted my code with `pylint`
- [x] I have added a one-line description of my change to the changelog
in
      `HISTORY.md`
- [x] This PR is ready to go