Move sentinel values for retrieve into a util function (#533)
## Description

<!-- Provide a brief description of the PR's purpose here. -->

To help reduce duplicate code across the archives, I moved the sentinel
values into a util function. While we have a lot of duplicate code
across the archives, we would like to take out some of the more mundane
things so that users can focus on what's important for archive
implementations.

## Status

- [x] I have read the guidelines in

[CONTRIBUTING.md](https://github.com/icaros-usc/pyribs/blob/master/CONTRIBUTING.md)
- [x] I have formatted my code using `yapf`
- [x] I have tested my code by running `pytest`
- [x] I have linted my code with `pylint`
- [x] I have added a one-line description of my change to the changelog
in
      `HISTORY.md`
- [x] This PR is ready to go