Move cqd_score into a separate function (#537)
## Description

<!-- Provide a brief description of the PR's purpose here. -->

Previously, `cqd_score` was a method on the archives. However, this led
to a lot of duplicate code, and `cqd_score` is not really a
functionality that is inherent to an archive. Thus, I have moved it to a
separate function `ribs.archives.cqd_score` and removed the methods from
the archives. I have also added an example usage in
`examples/cqd_score.py`. Finally, I updated the API
(backwards-incompatible) to always require passing in target points and
to not call the lower and upper bounds of the archive. This is because
not all archives have lower and upper bounds, and target points are not
always generated in a hyperrectangle within such bounds

## TODO

<!-- Notable points that this PR has either accomplished or will
accomplish. -->

- [x] Remove `cqd_score` method from archives
- [x] Add `ribs.archives.cqd_score` function in
`ribs/archives/_cqd_score.py`
- [x] Move `CQDScoreResult` into `ribs/archives/_cqd_score.py` as well
- [x] Add cqd_score to listing in `ribs/archives/__init__.py`
- [x] Separate the tests for cqd_score into
`tests/archives/cqd_score_test.py`
- [x] Add example usage for `cqd_score` in `examples/cqd_score.py`
- [x] Test the example in `tests/examples.sh`
- [x] Update example listings in `docs/examples.md` and
`docs/examples/cqd_score.md`

## Status

- [x] I have read the guidelines in

[CONTRIBUTING.md](https://github.com/icaros-usc/pyribs/blob/master/CONTRIBUTING.md)
- [x] I have formatted my code using `yapf`
- [x] I have tested my code by running `pytest`
- [x] I have linted my code with `pylint`
- [x] I have added a one-line description of my change to the changelog
in
      `HISTORY.md`
- [x] This PR is ready to go