Support solutions with non-1D shapes (#550)
## Description

<!-- Provide a brief description of the PR's purpose here. -->

Historically, pyribs has focused on supporting solutions with 1D shapes.
However, there are plenty of cases where 1D solutions are not ideal. For
example, in the case of strings as solutions, we want the solutions to
be scalars (shape `()`) so that the solutions can just be objects. In
the case of grids, solutions would need to be 2D. Seeing that the
ArrayStore can easily support such solutions, this PR removes the
restriction on 1D solutions and allows specifying tuple shapes.

It is true that not all emitters will be able to support such solutions.
For example, it does not make sense for EvolutionStrategyEmitter to
support string solutions. In short, our philosophy is that different
emitters will support different solutions, and the emitters will only
work with archives that have the correct solutions specified.

## TODO

<!-- Notable points that this PR has either accomplished or will
accomplish. -->

- [x] Update `solution_dim` in ArchiveBase
- [x] Update `solution_dim` in archives
- [x] Add test for various `solution_dim` in GridArchive
- [x] Update `solution_dim` in EmitterBase

## Status

- [x] I have read the guidelines in

[CONTRIBUTING.md](https://github.com/icaros-usc/pyribs/blob/master/CONTRIBUTING.md)
- [x] I have formatted my code using `yapf`
- [x] I have tested my code by running `pytest`
- [x] I have linted my code with `pylint`
- [x] I have added a one-line description of my change to the changelog
in
      `HISTORY.md`
- [x] This PR is ready to go