Add policy visualization to NS tutorial (#560)
## Description

<!-- Provide a brief description of the PR's purpose here. -->

Add visualization of policies to the NS tutorial. Previously, I did not
include it in the tutorial because the rollouts for rendering were
outputting different measures than the evaluations. See
https://github.com/adaptive-intelligent-robotics/Kheperax/issues/4

Update: For this PR, I included a version that simply plots the
positions of the robot during the rollouts based on the state
descriptors returned by the `evaluate` function. However, since the
robot's angle is not available in the descriptors, I did not include the
lasers, as they would point in the wrong direction.

## Status

- [x] I have read the guidelines in

[CONTRIBUTING.md](https://github.com/icaros-usc/pyribs/blob/master/CONTRIBUTING.md)
- [x] I have formatted my code using `yapf`
- [x] I have tested my code by running `pytest`
- [x] I have linted my code with `pylint`
- [x] I have added a one-line description of my change to the changelog
in
      `HISTORY.md`
- [x] This PR is ready to go