Migrate from yapf to ruff for formatting (#581)
## Description

<!-- Provide a brief description of the PR's purpose here. -->

Ruff seems to be a highly popular tool for Python nowadays. I believe
migrating to ruff would make pyribs just a bit more accessible for
contributors.

A followup PR would run ruff to format the entire codebase.

In addition to formatting, ruff also supports linting. I believe it
would also be great to migrate to ruff for linting since it is
significantly faster than pylint, but this will require a separate PR.

Here is the info for setting up ruff in pre-commit:
https://github.com/astral-sh/ruff-pre-commit

## TODO

<!-- Notable points that this PR has either accomplished or will
accomplish. -->

- [x] Update pre-commit
- [x] Update instructions in CONTRIBUTING
- [x] Remove yapf config in pyproject.toml
- [x] Increase pylint and isort line length to 88

## Status

- [x] I have read the guidelines in

[CONTRIBUTING.md](https://github.com/icaros-usc/pyribs/blob/master/CONTRIBUTING.md)
- [x] I have formatted my code using `yapf`
- [x] I have tested my code by running `pytest`
- [x] I have linted my code with `pylint`
- [x] I have added a one-line description of my change to the changelog
in
      `HISTORY.md`
- [x] This PR is ready to go