Update dependencies for lunar lander (#585)
## Description

<!-- Provide a brief description of the PR's purpose here. -->

Update the dependencies for the lunar lander tutorial and example.
Specifically, I loosened the gymnasium dependency to be `>=1.0.0` since
lunar lander seems fairly stable. I also removed the old
`importlib.reload` statement in the notebook as that no longer seems
necessary. Finally, I updated the installation command in `examples.sh`.

Tested in local Conda environments with Python 3.10/3.11, and on Colab.

## Status

- [x] I have read the guidelines in

[CONTRIBUTING.md](https://github.com/icaros-usc/pyribs/blob/master/CONTRIBUTING.md)
- [x] I have formatted my code using `isort` and `ruff`
- [x] I have tested my code by running `pytest`
- [x] I have linted my code with `pylint`
- [x] I have added a one-line description of my change to the changelog
in
      `HISTORY.md`
- [x] This PR is ready to go