Make all metrics start from values in archive stats (#586)
## Description

<!-- Provide a brief description of the PR's purpose here. -->

This PR makes all metrics start from the initial values defined in
`archive.stats` or `result_archive.stats`. Typically, these values are
all 0.0, and previously, we did set them to be 0.0. However, `0.0` is a
float, while the values in `archive.stats` are 0-D arrays. To make sure
the metrics store a list with values of the same type, we can just grab
the initial values from the stats object.

## Status

- [x] I have read the guidelines in

[CONTRIBUTING.md](https://github.com/icaros-usc/pyribs/blob/master/CONTRIBUTING.md)
- [x] I have formatted my code using `isort` and `ruff`
- [x] I have tested my code by running `pytest`
- [x] I have linted my code with `pylint`
- [x] I have added a one-line description of my change to the changelog
in
      `HISTORY.md`
- [x] This PR is ready to go